-- 将原本的 MessageEvents 重命名为 CommandEvents 只负责指令消息 -- 新增MessageEvents 负责 监听普通消息 -- 新增捕获 光佬key 功能
